requirements.txt
Faker==9.8.2
kafka-python==2.0.2
pymongo==3.12.1
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.3
psycopg2==2.8.6
